testing: don't use Expect() from the fake exec wrapper
Failing a ginkgo.Expect() from inside the Exec wrapper can result in a
panic in some far-flung part of the code, which if it doesn't get
caught correctly will result in a useless test failure.

Instead, just check the expectation the old-fashioned way, and if it
fails, call klog.Fatal(), which will immediately exit the test with
the relevant error message.

Signed-off-by: Dan Winship <danw@redhat.com>