Remove errors based on passing nil elements to MarshalPodAnnotations()
Currently passing nil elements to MarshalPodAnnotations will cause "<nil>"
to be added to the json fields where they are not appropriate.

This PR removes one specific example of attempting to pass a nil element
to MarshalPodAnnotations() and fixes an error in UnmarshalPodAnnotations()
which incorrectly handled a call to net.ParseIP() in which was expecting
an error in the case of an invalid IP address but in the docs returns nil
on invalid IP. This exposes the error of attempting to unmarshal an
invaild gatewayIP and add a test case to verify the error returned.

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>