hybrid-overlay: Remove ignoreLearn for external gw
We set this to `false` in 99% of cases.
The only time we do not is in the case of namespace update, and then
process and update for each pod in the ns with it set to true.

However, the pod update codepath will only run if the IP or MAC of the
pod has changed. In which case, the learn flow would also need to be
replaced.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>