scale: Enable parallel pod creation
This commit enables creating queued informers for pod events and
increases the number of event queues for handling events to 15.
At scale, the pod informer gets events more rapidly, but the handling
of these events is rather serialized. This will allow for multiple pod
events to be handled in parallel for a given namespace.

Scale tests were ran with this change.
At higher node scale (n > 150), the event handlers start contending if the numEventQueues is > 25.

Signed-off-by: Aniket Bhat anbhat@redhat.com