shared-gw: fix access to nodeport services from the node itself
In the shared gateway mode, the NodePort service is handled by the
OpenFlow flows configured on the OVS bridge in the host. These flows
act only on the packets coming in from outside of the node. If someone
on the node is trying to access the NodePort service, those packets
will not be processed by the OpenFlow flows, so we need to add iptable
rules that DNATs the NodePortIP:NodePort to ClusterServiceIP:Port.

    +---------------------------------------+
    |                 Node1                 |
    |             (10.8.51.51)              |
    |                                       |
    |                          +----------+ |
    |                          |  br-int  | |
    |   +--------+             +----+-----+ |
    |   |  Node  |             SNAT: NodeIP |
    |   |   IP   |                  |       |
    |   +--------+                  |       |
    | +-----------------------+     |       |
    | |        breth0         |   patch     |
    | |   NodePort OpenFlow   +-- cable     |
    | |         Flows         |             |
    | +-----------------------+             |
    |         +------+                      |
    +---------+ eth0 +----------------------+
              +------+

A DNAT rule in the OUTPUT chain (wrapped in ONV-KUBE-NODEPORT chain)
translates the packets to cluster ip and port. There are routes already
on the host to forward the packets to management port, and thus access
to NodePort backed service is achieved.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>