pkg/factory: increase QueuedHandler queue size
If two updates come in for the same pod in short succession, they will
block all other pod updates, since the function that muxes updates will
be blocked.

I saw this while running some simple benchmarks.

So, increase the queue size to 10, just to prevent most cases of
head-of-line blocking.

Signed-off-by: Casey Callendrello <cdc@redhat.com>