pkg/factory: fix deadlock
The QueuedInformer relies on RLock() always being able to make progress,
but this isn't always the case. Namely, if another goroutine calls
Lock(), further RLock() calls may fail even though the lock is free.

This causes a deadlock between the enqueuer and the consumer.

The fix is two-fold: drop the lock before enqueuing, and increase the
queue size, which should help efficiency anyways.

Signed-off-by: Casey Callendrello <cdc@redhat.com>