factory: remove unecessary locking for enqueueEevent
The only reason to lock the events array was to ensure that informer
shutdown worked correctly (mostly for unit tests), and that when
a nil event queue was found to stop distributing events to it.

Instead wait for all processEvent() goroutines for each event queue
to terminate before returning from Shutdown(), which handles stopping
event distribution. Just leave the event queues open as there is no
way to safely close now, because the shared informer may still be
delivering events to the factories handler before it notices that
the stop channel has been closed. This isn't a problem though, since
we only ever stop the informer on exit anyway.

Signed-off-by: Dan Williams <dcbw@redhat.com>