ovnkube.sh: ready_to_start_node() fails at scale
the entrypoint script for various ovn k8s containers call
ready_to_start_node() to see if they can continue starting the process
for that container. This function makes a call to ovn-nbctl CLI and this
call fails, timesout, at scale with 100s of nodes in the cluster.

The call to ovn-nbctl is not required. All we need to check in this
function is to see if ovnkube-db kubernetes service has ovnkube-db
endpoints populated. The logic that populates these endpoints would
have already ensured that the OVN DBs are available at the said
endpoints, so there is no need to check again.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>