hybrid-overlay: fix host route to hybrid overlay subnets
Don't use direct byte manipulation of the node subnet, otherwise
it's changed for all subsequent uses and that causes the host route
via the management port for the hybrid overlay cluster subnet to
get a .6 address instead of the expected .3.

In addition the testcases didn't have the right hybrid overlay
CLI options to enable the right code checks.

Signed-off-by: Dan Williams <dcbw@redhat.com>