hybrid-overlay: Fix pod logging and filter node name in event handlers
This commit fixes the logging for pod add/update/delete events to just
emit the pod name and not to attempt to print the entire pod object as
it makes the logs impossible to read.

It also adjusts getPodDetails to not require a node name.
This filtering is done in the resourceEventHandlerFuncs instead.
This is required to avoid cases where getPodDetails would fail in
the add/update function as the pod wasn't annotated yet, and would
then call delete on error, resulting in another error again due to
lack of annotations, making the logs very chatty.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>