topology changes to enable Node-Local Services Access in Shared GW Mode
In the shared gateway mode, the current OVN Logical Topology does not
support hostNetwork pods from accessing a kubernetes ClusterIP service
endpoint which happens to be the node itself. Such traffic gets dropped.

In this mode, we share the IP address of the node to SNAT the traffic
going out of the OVN Logical network to the physical network. The OVN
Gateway router does this SNATing. When a packet with destination IP set
to the Node IP arrives on the `ingress` pipeline of the Gateway Router,
it expects the packet to be one of these things:
  (1) a packet for NodePort service and it DNATs it to one of the
     backend IPs
  (2) a reply packet for an already established (or in the process of)
     connection
If it sees a packet from the OVN logical topology towards the Node IP,
then it gets dropped since it doesn’t fit (1) or (2) above.

This commit provides a path for such traffic by using OVN features of
Policy-Based routing and one-to-one DNAT and SNAT on the distributed
gateway port.

The google doc version of the design document is here:
https://docs.google.com/document/d/\
 1DfbcAezbZD4Ipswz6UOn30S6B6f6eoiLkTCVmTa43Vk/edit?usp=sharing

The Markup version of the design document along with the diagrams is in
docs/design/shared_gw_dgp.md

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>