retry addNode() again in NodeUpdate after it failed in NodeAdd handler
Say we failed in addNode() for worker node node1. Now, it doesn't get
cached in oc.logicalSwitchCache. We don't try to call addNode() again
on NodeUpdates if addNode() failed in NodeAdd.

With this state for node1, all the pods that gets scheduled on that node
will not proceed since the addLogicalSwitchPort() continues to wait for
the node to appear in logicalSwitchCache. It times out with the error
below
------8<---------8<-----
timed out waiting for logical switch "node1" subnet: timed out waiting\
 for the condition
------8<---------8<-----

all such pods will be in CrashLoopBackOff state.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>