test: Check network readiness before connection tests
This commit asks our tests to poll for the presence of annotations
on the pods that are created as some signal that the network is,
almost, ready.

The OVN/HO master are responsible for writing these annotations.
The nodes wait on the presence of theses anootations before processing
can continue.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>