Fix order of pods, namespaces, services and endpoints
On a restart of ovnkube master due to leader election change
(normal) and perhaps loading and processing all the existing pods.

E0609 18:43:27.194389 1 ovn.go:396] timeout waiting for namespace event

Occured. We process services and endpoints before namespaces, which is
the wrong order because pods (which are first) depend on Namespaces.
So if there are a lot of services & endpoints their initial processing
will block the namespaces which will block and fail pods.

SDN-1037 - Processing services and endpoints incorrectly happens before
namespaces.
https://issues.redhat.com/browse/SDN-1037

Signed-off-by: Phil Cameron <pcameron@redhat.com>