Add sleep timer to HO test
There's a race between when the namespace is updated and the new pod
that gets created right after. The problem is nsInfo in the ovn-k8s
backend may not be updated (has to wait for lock) by the time the new
pod gets created. A 1 second sleep fixes the issue.

Signed-off-by: Tim Rozet <trozet@redhat.com>