expose gateway mode for ovnkube-master process
a change was made to OVN logical topology to include distributed gateway
port to provide access to node local services from hostnetwork pods.
this is needed only for the shared gateway mode. there is a check in
place in master startup code to create these ports for shared gateway
mode alone. however, the default value of the gateway mode is 'shared'
and the yamls generated didn't have a way to change this to 'local'
mode and as a result the topology change was happening for the local
gateway mode too.

Fixes issues #1446

Co-authored-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>