Support go-ovn bindings for addLogicalPort and e2e timestamps
This change uses go-ovn client to do logical port additions and also
for the timestamp metrics. There are additional validations needed on
the unit-tests including verifying the mock db contents which will be
added in a subsequent commits. Also pending are changes to pass the
correct hostname parameter for SSL certs to work correctly in the conf
file.

Signed-off-by: Aniket Bhat <anbhat@redhat.com>