trivial: incorrect prefix used for the port connecting switch to router
during the code review of edb24e6a7114 (topology changes to enable
Node-Local Services Access in Shared GW Mode) it was decided to use
the prefix of rtos- and stor- for the ports connecting node_local_switch
and the ovn_cluster_router. however, for the switch port we are still
using a different convention.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>