use kubernets upstream ginkgo e2e script
Kubernetes upstream has a helper script to run ginkgo,
that allows to can run tests in parallel, that will improve
the time to run the jobs by several orders of magnitude.
From 50 mins to 7 mins to run the whole sig-network tests

This way we don't neet to shard the tests in several jobs.

It also allows to configure retrying flake tests,
We are already retrying PRs, this at least will
give signal on totally broken tests, if a test
is not working after X times, that's something
we should investigate.

Signed-off-by: Antonio Ojea <aojea@redhat.com>