use ginkgo parallel
ginkgo can run tests in parallel, that will improve
the time to run the jobs by several orders of magnitude.
From 50 mins to 7 mins to run the whole sig-network tests

This way we don't neet to shard the tests in several jobs.

It is configured to run 10 tests in parallel

Signed-off-by: Antonio Ojea <aojea@redhat.com>