Make output of testing functions mirror the real functions
Currently the output of GetIPv4HashName() and GetIPv6HashName() in the
fakeAddressSet implementation do not mirror the functionality of the
GetIPv4HashName() and GetIPv6HashName() in the OvnAddressSet implementation
of the AddressSet interface. This will cause functions that call GetIPvXHashName()
to behave differently under test then under normal operating conditions.

This patch corrects that and makes the output the same

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>