Simplify SubnetAllocator ranges, fix hybrid overlay problem
The configuration always specifies networks as base CIDR plus
HostSubnetLength, but the allocator (for historical reasons) required
the HostSubnetLength to be converted to its inverse, which then
requires the caller to unnecessarily take into account whether it's an
IPv4 or IPv6 range (which the hybrid overlay code was not doing).

Fix this by just making SubnetAllocator.AddNetworkRange() take a
HostSubnetLength instead, and also make it an int rather than uint32,
for ease of use, and then update everything for that.

Signed-off-by: Dan Winship <danwinship@redhat.com>