ovn: don't log spurious warnings/errors on node creation
After a new Node is created, it will take a while before the
ovn-kubernetes process on that node annotates the node with the
information the master needs. The master should not repeatedly log
"errors" about the annotations not being set, because this is
completely expected.

Also, make ParseNodeManagementPortMACAddress() consistent with all of
the other annotation-parsing methods, and have it return an error when
the annotation is unset. This requires updating one caller, which was
confused anyway because it thought it was checking if the annotation
was removed, but we never remove the annotation after setting it;
really it was only checking if the annotation had not been set yet.

Signed-off-by: Dan Winship <danwinship@redhat.com>