enable go race-detector for testing
The race detector is integrated with the go tool chain. When
the -race command-line flag is set, the compiler instruments
all memory accesses with code that records when and how the
memory was accessed, while the runtime library watches for
unsynchronized accesses to shared variables. When such
"racy" behavior is detected, a warning is printed

However, race-enabled binaries can use ten times the CPU and
memory, so it is easy to fail tests that expect asynchronous
assertions

https://blog.golang.org/race-detector

Signed-off-by: Antonio Ojea <aojea@redhat.com>