remove the logic that retries on 'Connection Refused' error
The code in runOVNretry() is from early days of OVN Kubernetes project.
The commit that introduced it says

   "ovnkube can't start until the ovn-ndb is running. This change goes
    into a loop waiting until the daemon is running"

This isn't the case anymore. We have all the logic in the entrypoint
script to ensure that the OVN NB and SB DB are up and running before
any of the OVN K8s components come up.

Consider this scenario. The ovn-northd daemon restarted. The metrics
for ovn-northd are collected from ovnkube-master daemon. The call to
ovn-northd (ovn-appctl -t /var/run/openvswitch/ovn-northd.67.ctl status)
got stuck in runOVNretry() for about 200 tries * 2 seconds = 400
seconds, and all the prometheus call to retrieve ovn metrics hung. This
went on until ovn-northd came up. The right thing to do here should
have been to return with error.

Whether to retry or not should be left to callers using wait.Poll*()
functions.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>