ovn: fix setting addresses on logical switch ports
Addresses must be a space-separated string sent as a single value.
The slice passed to LSPSetAddress() turns into an OvsSet in which
has special semantics for the second and later elements and causes
OVSDB to insert a newline between the MAC and IPs, leading to
northd complaints that the address is malformed.

Signed-off-by: Dan Williams <dcbw@redhat.com>