Remove native Windows support
Remove the ability to run ovn-kubernetes natively on Windows. This
hasn't worked for a while as ovnkube has required OVN 2.11 for the
past couple months, which isn't available on Windows that we know
of.

Note that this does not affect the Windows hybrid-overlay code and
binary. That is still supported.

Signed-off-by: Dan Williams <dcbw@redhat.com>