Port need not be specified for a protocol in a network policy
K8s network policy allows to specify the protocol, but if the port
is not specified, it is taken to mean all the ports for that
protocol

"If specified, the port on the given protocol.  This can either be a
numerical or named port on a pod.  If this field is not provided, this
matches all port names and numbers. If present, only traffic on the
specified protocol AND port will be matched.
+optional"