Fixes shared gateway mode checks (#1535)
* Fixes shared gateway mode checks

The current code for detecting the NIC with shared gateway mode assumes
the naming format of the bridge is "br<nic name>" or that an external id
with name "bridge-uplink" is added to the bridge. With an OpenShift
deployment we use a bridge name of "br-ex" and the only external ID is
the NetworkManager UUID of the connection.

This patch adds an alternative way of finding the NIC port, by checking
if the port type is "system". If using Network Manager, NM will always
configure a Type on the port, and it will use system for physical ports.

Additionally, checks for ofport were using "--if-exists" which will not
return an error if the ofport does not exist for that port.

Signed-off-by: Tim Rozet <trozet@redhat.com>