bump kubernetes version for e2e tests
refactor and adapt e2e cp tests

- adapt the code to the 1.18 api
- adapt the tests to run offline
- support IPv6 in all tests
- remove kubectl dependency
- remove deadlocks

The e2e tests were using a goroutine to test that connectivity
on disruptive scenarios.
This complicates the test and exposes tehe developer to deadlocks
problem, instead we can send a pod in the background to test
the connectivity during a specified time period and check the result
once it finish.

Signed-off-by: Antonio Ojea <aojea@redhat.com>