Add a retry mechanism for setting node annotations
During node addition if setting annotations fails because API server
is down, currently we don't have a mechanism to retry. We add the node
to a datastore for failed nodes. But the node will only be retried
in the updateFunc when it truly is updated. Barring any updates, the node
will never get an annotation. This retry will only account for occasional
blips in communication to the API server. Any outage > 1s will result in a
failure annotating the node.

Signed-off-by: Aniket Bhat <anbhat@redhat.com>