Switch to correct egress IP implementation for shared gw mode
The initial implementation of egress IP in shared gateway mode
was dependent on having the bug: https://bugzilla.redhat.com/show_bug.cgi?id=1861294
resolved. This is now the case and the fix is in the latest Fedora 31 release.
This commit thus switches to the correct shared gateway mode implementation.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>