factory: Don't return errors from Remove*Handler
The only way the Remove*Handler functions can fail is if you try to
remove a handler twice. That is a programmer error that should be
fixed, but there's nothing the caller can do in response to it, and as
far as the caller is concerned, the remove succeeded anyway. So just
log an error in this case and don't return any error to the caller.

Signed-off-by: Dan Winship <danwinship@redhat.com>