factory: Don't return errors from Add*Handler
AddHandler can only fail if it is given an invalid LabelSelector. But
most of the callers always pass nil for the LabelSelector, and so can
never actually fail (and yet need to have code to check for failure
anyway). In fact, most of the filtered handler can never fail either,
because they're passing LabelSelectors that have already been
validated to be correct.

To get rid of the unnecessary error checking, change the filtered
handler functions to take a (parsed/validated) labels.Selector rather
than an (unparsed/unvalidated) metav1.LabelSelector, to push the error
checking on that back to the callers (if even needed). Then the actual
adding-a-handler part becomes non-failable.

Signed-off-by: Dan Winship <danwinship@redhat.com>