factory: Adding a handler of unknown type is a programmer error
If this happens, it means this is a developer build with some buggy or
incomplete feature. Don't return an error to the caller, just die and
let the developer fix it.

Also, we don't need to check if the type is known when removing a
handler, because the handler couldn't exist if the type was unknown.

Signed-off-by: Dan Winship <danwinship@redhat.com>