Ignore add pod errors for network policy selector funcs
Currently the handlePeerPodSelectorAddUpdate function looks
for add pod events, which can never be successful. This is
because the policy code relies on finding an IP in the
annotation or pod status, which is not set until after
addLogicalPort is complete and an update event comes.

For handleLocalPodSelector the add event would require the port to
already be in the cache. This is an unlikely scenario, because
addLogicalPort will take a lot longer to get the port into cache by the
time this function has already checked the cache.

This patch ignores errors fot not being able to find port information
during pod Add, to cut down on the number of false errors we are seeing
in logs.

Signed-off-by: Tim Rozet <trozet@redhat.com>