Pod Add: handle existing annotation errors
Currently if a pod is added with an existing annotation, we configure it
in OVN, but never try to allocate the IPs in our allocator. Potentially
this could result in another pod being created and getting the same IP
address since it is untracked by the allocator. This patch ensures the
allocator is able to get the address, otherwise clear the existing
annotation.

Also, refactor the IP allocation to make things clearer. The order of
operations is:
 1. Attempt to find IP via annotation, ensure allocation
 2. Attempt to find IP via OVN port, ensure allocation
 3. Assign and allocate new IP address

For all of these the end result is allocation, so always releaseIPs.

Signed-off-by: Tim Rozet <trozet@redhat.com>