don't process objects set for deletion
The first delete request on an object with finalizers sets
a value for the metadata.deletionTimestamp field but does not
delete it.

Our informer implementation, on updates, was checking for the
DeletionTimestamp field in the old object instead of the new object,
thus we were processing events for objects that are set for deletion.

After checking the kubernetes code base, I've have observed that
objects can be set for deletion on Add events, and controllers
are filtering on add events too.

We should not process any object that is set for deletion, i.e. that
has the DeletionTimestamp field != nil

Signed-off-by: Antonio Ojea <aojea@redhat.com>