Annotate pods early
Since we can rely on CNI to wait until the OpenFlow flows are present,
we can safely try to annotate the pod first on pod create. Most of the
time the add_logical_port function fails because the pod has been
deleted in k8s, so by the time we fail to annotate the pod, all of the
OVN configuration has already been done.

This patch annotates just after the IP allocation is done. With these
changes IP allocation is only rolled back if annotation fails. The
reason for this is that we can consider the IP allocated and IP
annotated on the pod to be persistent throughout the lifecycle of the
pod. Failures to add logical port still result in a retry, but the
annotation is not set again, and the expectation is that allocation is
already done, but that is also ensured.

Additionally, this change includes modifying the behavior of locking the
namespace during pod add. The namespace must exist at some point during
the pod add for various reasons. This patch modifies addPodToNamespace
to wait for the namespace to exist, and all subsequent calls to lock
namespace in this code path should be changed from
waitForNamespaceLocked to getNamespaceLocked (future patch).

Signed-off-by: Tim Rozet <trozet@redhat.com>