Remove rolling back OVN config
In addLogicalPort there is no reason to roll back OVN configuration.
Since the configuration should be idempotent, repeating the
configuration should be fine. The final configuration now of LSP port
security is used to signal CNI is complete, and if that operation
suceeds addLogicalPort succeeds. If it fails, then it never configured
it, so there is nothing to roll back.

Signed-off-by: Tim Rozet <trozet@redhat.com>