Handle testing race condition in egressip_test.go
The test `should perform re-assignment of EgressIPs` in `egressip_test.go` sometimes panics
as `getEgressIP` (which is used to retrieve that the egress IP status during the asynchronous update)
can return an empty status. This is because `WatchEgressIP` on UPDATE resets the
status field to an empty status before performing the re-assignment. `deleteEgressIP` and `addEgressIP`
both use a mutex to assure correct handling during the DELETE / ADD, but during the UPDATE
there is a small window where the status field can be empty. This solution does a small work-around
in the test to avoid that panic, but does not cause the test to fail and subsequently continue trying.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>