Re-work and fix ovn-nbctl transactions for egress IP setup
Tag egress IP related OVN transactions with external_id

This patch deals with avoiding inconsistent state when a pod matches multiple
egress IPs. Before this patch: the lr-policy-add / lr-nat-add commands failed once the second
egress IP setup happened. This wasn't a problem in itself, but if that egress IP was later
deleted: it then proceeded to delete the egress setup for the first egress IP. It goes like this:

1) EgressIP-1 is created matching pod X, setup happens correctly
2) EgressIP-2 is created matching pod X, setup fails because that exact same setup (nat and policy rules) already exists
3) EgressIP-2 is deleted, deleting EgressIP-1's setup.
4) OVN state is now screwed up and pod egress functionality will no longer work even though EgressIP-1 still exists.

To avoid this we have the possibility to tag the policy and nat objects with the external_id,
which in this case is the EgressIP.Name. As such we will only try deleting whatever setup we find
for that particular object. The OVN API for doing this is not as straightforward though, and we use
an extra OVN transaction for the delete, but we win in correctness.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>

Expanding on the docstring for policyAlreadyExistsMsg

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>

Simplify logical router policy filter for egress IP

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>

Check if nat/lr-policy exists before creating it

As to avoid duplicate objects in the OVN database, check that
the NAT / logical router policy object which is about to be created
does not already exist

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>