Use action flood intead of normal for shared gw bridge
Today we use conntrack to make decisions on how to steer traffic between
the host and OVN on the shared gw bridge (breth0 in KIND cases). This
necessary because we share the same MAC/IP between OVN and the host.
However, some traffic cannot be conntracked (like IPv6 ND) so therefore
it will hit the default rule which currently is NORMAL action.

However, NORMAL action is not sufficient because that will end up
choosing whichever entry is in the FDB, and that is undeterministic
because both ports are sharing the same MAC. This patch changes the
behavior to use FLOOD so that unknown packets are sent to both the host
and OVN.

Signed-off-by: Tim Rozet <trozet@redhat.com>