Make sure help message is always printed for kind.sh
Moved installation of j2cli via PIP after parsing the arguments.
Parsing command line arguments should be the first statement in the script so that if someone is just providing
the "-h" flag we print out the help  without doing any check. By moving the above statement as mentioned
we have ensured "-h" flag works even if say pip is not installed in the system.

Signed-off-by: Balaji Varadaraju <bvaradar@redhat.com>