Ensure Raft membership is correct for OVN DBs
There are two scenarios where a stale member may be leftover in the Raft
membership:

1. OVN DB pod goes down, database is wiped, comes back up with same
address but gets a new server ID in Raft

2. DB node is removed from cluster and replaced with a new one with a
different address

For 1, each ovnkube-node can check its current ID and look for any
leftover members with its same address if the DB is on the same node
as the ovnkube-node instance.

For 2, ovnkube-nodes with co-located DBs will compare the known
addresses passed to it for other DB members, and kick out any stale
members.

Signed-off-by: Tim Rozet <trozet@redhat.com>