Fixes raft membership configuration
The current code always assumed that if ovnkube-db-0 was coming up and
the DB was gone that it should create a new cluster. This would result
in errors like:

append_request RPC failed: wrong cluster a8fc (expected 9052)

This commit modifies the behavior so that we check to see if there is a
current cluster before asking ovnkube-db-0 to create one.

Signed-off-by: Tim Rozet <trozet@redhat.com>