Make sure lr-nat-add doesn't fail when adding rows
The lr-nat-add would occasionally fail in the gw initialization code
causing the nodes to not come up. The --may-exist semantics work on
the tuple as a whole and if a given row is added again with a different
external IP in case of snat and different logical IP in case of dnat or
dnat_and_snat, the add would fail. So this change makes sure we delete the nat
entry first before trying to add it.

Signed-off-by: Aniket Bhat <anbhat@redhat.com>