Revert "Fix default allow mgmt ACL using conntrack"
This reverts commit fd4758701cd61e0f69e21ef5a96ab5d91f704ef0.

We found that this causes problems with network policy ACLs when using
multi-node setups where the network policy ACL only targets a portgroup
of the destination, and in effect only places the ACL there with
allow-related. This causes return traffic destined to another node to be
dropped, because that node does not have any allow-related ACL, and thus
never able to determine via CT if this is reply traffic.

Signed-off-by: Tim Rozet <trozet@redhat.com>