Handle egress IP assignment for node IPs
When an egress IP is specified equaling a node's IP address, we
can end up in the situation where the assignment procedure assigns the
egress IP to another cluster node, and thus missing the node actually hosting
that IP address. This patch thus fixes this by invoking an error during the
assignment procedure and triggering an event notifying him/her of this
unsupported request.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>